Sis222/SIS2/requirements.txt
selenium==4.15.0
pandas==2.1.3
apache-airflow==2.7.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
chromedriver-autoinstaller
numpy==1.26.4